# docs/openapi.yaml
openapi: 3.0.3
info:
  title: CM FIM System API
  version: "1.0"
  description: >
    Fleet Incident & Insurance Claim Management System (JBM) — the
    protected JSON API under /api. Every route requires a valid session
    cookie (cm_fim_session); there is no separate API-key auth. Hand-
    maintained against the actual route handlers in app/api/** rather
    than generated, matching this project's "no unnecessary dependencies"
    convention (docs/PAYMENTS.md, docs/ESCALATIONS.md, ...) — keep it in
    sync when routes change. See docs/SCOPE.md for the domain model and
    docs/AUTH.md for the session/RBAC mechanics this spec assumes.
servers:
  - url: /api
security:
  - sessionCookie: []
...
paths:
...
  /incidents:
    get:
      tags: [Incidents]
      summary: >-
        List incidents (DEPOT_MANAGER sees only their own depot), with
        M21's richer filters
      parameters:
        - name: status
          in: query
          schema: { type: string, enum: [OPEN, CLOSED] }
        - name: severity
          in: query
          schema: { $ref: "#/components/schemas/IncidentSeverity" }
        - name: incidentType
          in: query
          schema: { $ref: "#/components/schemas/IncidentType" }
        - name: depotId
          in: query
          schema: { type: string, format: uuid }
        - name: dateFrom
          in: query
          schema: { type: string, format: date }
        - name: dateTo
          in: query
          schema: { type: string, format: date }
      responses:
        "200":
          {
            description: OK,
            content:
              {
                application/json:
                  {
                    schema:
                      {
                        type: array,
                        items: { $ref: "#/components/schemas/Incident" },
                      },
                  },
              },
          }
    post:
      tags: [Incidents]
      summary: Report an incident (ORG_ADMIN, DEPOT_MANAGER) — BR-01
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [vehicleId, incidentDateTime, incidentType, description]
              properties:
                vehicleId: { type: string, format: uuid }
                driverId: { type: string, format: uuid }
                incidentDateTime: { type: string, format: date-time }
                incidentType: { $ref: "#/components/schemas/IncidentType" }
                severity: { $ref: "#/components/schemas/IncidentSeverity" }
                locationAddress: { type: string, maxLength: 500 }
                locationLat: { type: number }
                locationLng: { type: number }
                description: { type: string, minLength: 1, maxLength: 4000 }
                reportedVia:
                  {
                    type: string,
                    enum: [WEB, WHATSAPP, MOBILE_APP, PHONE, OTHER],
                  }
                injuries:
                  {
                    type: string,
                    maxLength: 2000,
                    description: "M21 — the Overview tab's Driver section",
                  }
                thirdPartyInvolved: { type: boolean, description: "M21" }
      responses:
        "201":
          {
            description: Created — auto-instantiates this org's INCIDENT-type TAT stages,
          }
        "400": { $ref: "#/components/responses/ValidationError" }
        "403": { $ref: "#/components/responses/Forbidden" }
...
  /tat/dashboard:
    get:
      tags: [Dashboards]
      summary: >-
        A live board of every active (IN_PROGRESS/ON_HOLD) stage across
        incidents and claims — not a history report, see docs/TAT.md's
        M23 section
      parameters:
        - name: depotId
          in: query
          schema: { type: string, format: uuid }
        - name: caseType
          in: query
          description: The real CaseType enum, read off each stage's TatStageTemplate.
          schema: { $ref: "#/components/schemas/CaseType" }
        - name: breachedOnly
          in: query
          schema: { type: boolean }
      responses:
        "200": { description: OK }

  # --- MIS Reports (M24) ----------------------------------------------------
...
  /fleet:
    get:
      tags: [Dashboards]
      summary: >-
        Fleet-wide KPIs plus a filterable vehicle list — see
        docs/DASHBOARDS.md's M25 section
      parameters:
        - name: depotId
          in: query
          schema: { type: string, format: uuid }
        - name: status
          in: query
          schema: { $ref: "#/components/schemas/VehicleStatus" }
        - name: hasOpenIncidents
          in: query
          schema: { type: boolean }
        - name: hasOpenClaims
          in: query
          schema: { type: boolean }
      responses:
        "200":
          {
            description: "OK — { kpis, vehicles }",
          }

  # --- My Work (M26) ---------------------------------------------------------
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: cm_fim_session
      description: >
        An encrypted JWT (jose/HS256) wrapping {sessionId}. The real
        authorization check is the DB-backed Session row this points at
        (lib/dal.ts's verifySession()) — see docs/AUTH.md. Minted at
        login; there is no separate token-issuing endpoint in this API.
...
  responses:
...
    Forbidden:
      description: Session exists but the role isn't allowed (403)
      content:
        { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...
    ValidationError:
      description: Request body failed Zod validation (400)
      content:
        { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
...
    IncidentType:
      type: string
      enum:
        [
          ACCIDENT,
          THEFT,
          FIRE,
          BREAKDOWN,
          NATURAL_DISASTER,
          THIRD_PARTY_DAMAGE,
          OTHER,
        ]
    IncidentSeverity:
      type: string
      enum: [LOW, MEDIUM, HIGH, CRITICAL]
    VehicleStatus:
      type: string
      enum: [ACTIVE, INACTIVE, SOLD, SCRAPPED]
    Incident:
      type: object
      properties:
        id: { type: string, format: uuid }
        organizationId: { type: string, format: uuid }
        incidentNumber: { type: string, example: "INC-2026-000001" }
        vehicleId: { type: string, format: uuid }
        driverId: { type: string, format: uuid, nullable: true }
        depotId: { type: string, format: uuid }
        incidentDateTime: { type: string, format: date-time }
        incidentType: { $ref: "#/components/schemas/IncidentType" }
        severity: { $ref: "#/components/schemas/IncidentSeverity" }
        description: { type: string }
        status: { type: string, enum: [OPEN, CLOSED] }
        reportedById: { type: string, format: uuid }
        createdAt: { type: string, format: date-time }
        injuries:
          {
            type: string,
            nullable: true,
            description: "M21 — the Overview tab's Driver section",
          }
        thirdPartyInvolved: { type: boolean, nullable: true, description: "M21" }
...
    CaseType:
      type: string
      enum:
        [
          INCIDENT,
          INSURANCE_CLAIM,
          WARRANTY_CLAIM,
          MAINTENANCE_CLAIM,
          OPERATIONAL_CLAIM,
          THIRD_PARTY_RECOVERY_CLAIM,
          MIXED_CLAIM,
        ]
...